==> modelagem/01 - FirstGlance_and_Data_Preparation.ipynb <==
import pandas as pd 

link = 'https://s3-sa-east-1.amazonaws.com/ckan.saude.gov.br/SRAG/2021/INFLUD21-27-09-2021.csv'
df = pd.read_csv(link, sep=";", low_memory=False)

df.shape

name='INFLUD21-27-09-2021'

df.head()

columns = ['DT_NOTIFIC', 
        'DT_SIN_PRI', 
        'CS_SEXO',
        'CS_GESTANT',
        'CS_RACA',
        'NOSOCOMIAL',
        'FEBRE',
        'TOSSE',
        'GARGANTA',
        'DISPNEIA',
        'DESC_RESP',
        'SATURACAO',
        'DIARREIA',
        'VOMITO',
        'PUERPERA',
        'FATOR_RISC',
        'CARDIOPATI',
        'HEMATOLOGI',
        'SIND_DOWN',
        'HEPATICA',
        'ASMA',
        'DIABETES',
        'NEUROLOGIC',
        'PNEUMOPATI',
        'IMUNODEPRE',
        'RENAL',
        'OBESIDADE',
        'VACINA',
        'ANTIVIRAL',
        'DT_ENTUTI',  
        'DT_SAIDUTI', 
        'SUPORT_VEN',
        'CLASSI_FIN', 
        'CLASSI_OUT', 
        'EVOLUCAO', 
        'DOR_ABD',
        'FADIGA',
        'PERD_OLFT',
        'PERD_PALA',
        'VACINA_COV',
        'DOSE_1_COV',
        'DOSE_2_COV'
]
df_sub = df[columns]

df_sub.shape

This study has the objective to predict the use or artificial ventilators in patients of acute respiratory syndrome 

Column - 'SUPORT_VEN'

This column has the following items: 

* 1-Sim, invasivo 
* 2-Sim, não invasivo 
* 3-Não
* 9-Ignorado

Missing data and values iqual 9 has to be filtered out
'''

from google.colab import drive
drive.mount('/content/drive')

df_filtered = df_sub[df_sub['SUPORT_VEN'].notnull()]
df_filtered.shape

df_filtered = df_filtered[df_filtered['SUPORT_VEN']!= 9]
df_filtered.shape

df_filtered.to_pickle('INFLUD21-27-09-2021-SupportVent_01.pkl')
