==> modelagem/02 - Simple filtering missing values.ipynb <==
import pandas as pd 
from google.colab import drive
drive.mount('/content/drive')

df = pd.read_pickle('/content/drive/MyDrive/Grupo Facens DataScience/TCC/Grupo 1/Codes/Arquivos/INFLUD21-27-09-2021-SupportVent_01.pkl')
df.shape

teste = ['DT_NOTIFIC',
         'DT_SIN_PRI', 
         'CS_GESTANT', 
         'PUERPERA',   
         'CARDIOPATI',
         'HEMATOLOGI',
         'SIND_DOWN',
         'HEPATICA',
         'ASMA',
         'DIABETES',
         'NEUROLOGIC',
         'PNEUMOPATI',
         'IMUNODEPRE',
         'RENAL',
         'OBESIDADE',
         'DT_ENTUTI',  
         'DT_SAIDUTI', 
         'CLASSI_FIN', 
         'CLASSI_OUT', 
         'EVOLUCAO',       
         'DOSE_1_COV',
         'DOSE_2_COV'] 
len(teste)

cols = ['CS_SEXO',
        'CS_RACA',
        'NOSOCOMIAL',
        'FATOR_RISC',
        'CARDIOPATI',
        'HEMATOLOGI',
        'SIND_DOWN',
        'HEPATICA',
        'ASMA',
        'DIABETES',
        'NEUROLOGIC',
        'PNEUMOPATI',
        'IMUNODEPRE',
        'RENAL',
        'OBESIDADE',
        'FEBRE',
        'TOSSE',
        'GARGANTA',
        'DISPNEIA',
        'DESC_RESP',
        'SATURACAO',
        'DIARREIA',
        'VOMITO',
        'DOR_ABD',
        'FADIGA',
        'PERD_OLFT',
        'PERD_PALA', 
        'VACINA',
        'ANTIVIRAL',
        'SUPORT_VEN',
        'VACINA_COV']

#Filtering All columns in wich missing datas should be excluded. 
df_filter = df.copy()

for col in cols: 
    df_filter = df_filter[df_filter[col].notnull()]


#checking remaining rows: 
print(df_filter.shape)

df.shape[0] - df_filter.shape[0]

df_filter.shape

df_filter.to_pickle('INFLUD21-27-09-2021-NoMissing.pkl')
