==> modelagem/03 - Final_Data_Treatment.ipynb <==
import pandas as pd 
import numpy as np
from google.colab import drive
from datetime import datetime
drive.mount('/content/drive')

df = pd.read_pickle('/content/drive/MyDrive/Grupo Facens DataScience/TCC/Grupo 1/Codes/Arquivos/INFLUD21-27-09-2021-NoMissing.pkl')

df.head()

df.dtypes

ColumnsObject =  ['CS_GESTANT', 
         'PUERPERA',  
         'CARDIOPATI',
         'HEMATOLOGI',
         'SIND_DOWN',
         'HEPATICA',
         'ASMA',
         'DIABETES',
         'NEUROLOGIC',
         'PNEUMOPATI',
         'IMUNODEPRE',
         'RENAL',
         'OBESIDADE',   
         'DOSE_1_COV', 
         'DOSE_2_COV',
        'CS_RACA',
        'NOSOCOMIAL',
        'FATOR_RISC',
        'CARDIOPATI',
        'HEMATOLOGI',
        'SIND_DOWN',
        'HEPATICA',
        'ASMA',
        'DIABETES',
        'NEUROLOGIC',
        'PNEUMOPATI',
        'IMUNODEPRE',
        'RENAL',
        'OBESIDADE',
        'FEBRE',
        'TOSSE',
        'GARGANTA',
        'DISPNEIA',
        'DESC_RESP',
        'SATURACAO',
        'DIARREIA',
        'VOMITO',
        'DOR_ABD',
        'FADIGA',
        'PERD_OLFT',
        'PERD_PALA', 
        'VACINA',
        'ANTIVIRAL',
        'SUPORT_VEN',
        'VACINA_COV']



for columns in ColumnsObject:
    df[columns]=df[columns].astype('object')

df.dtypes

# 'PUERPERA'

#remove todas as linhas onde Puerpera possui valores vazios
df = df[df['PUERPERA'].isnull()!=True]
df.shape

#Valores = 1, algum se refere a homem? / Valores = 1, Algum aparece como gestante? 
print("Valores de Puerpera", df['PUERPERA'].unique())
print("Valores de Puerpera para pacientes do sexo Masculino",df.loc[df['CS_SEXO']=='M','PUERPERA'].unique())
print("Valores de Puerpera para pacientes gestante",df.loc[df['CS_SEXO']=='M','CS_GESTANT'].unique())

# 'DT_NOTIFIC' & 'DT_SIN_PRI'

#Transformar colunas'DT_NOTIFIC','DT_SIN_PRI'  para formato datetime 
df['DT_NOTIFIC'] = pd.to_datetime(df['DT_NOTIFIC'])
df['DT_SIN_PRI'] = pd.to_datetime(df['DT_SIN_PRI'])

#Não utilizar pois o dado não passa confiança devido a existir datas de notificação futura, cerca de 110k de registros

# 'CS_GESTANT'

## 'CS_GESTANT' Filtrar as linhas com valor 0 / 
print("Numero de linhas em 'CS_GESTANT' com valor igual a 0: ",df[df['CS_GESTANT']==0].shape[0])
df = df[df['CS_GESTANT']!=0] 

#/ Manter os valores 1,2,3,4 e Transformar para 1(sim)
df.loc[df['CS_GESTANT'].isin([1,2,3,4]),'CS_GESTANT'] = 1

#Valores 5 e 6 transformar para 0 
df.loc[df['CS_GESTANT'].isin([5,6]),'CS_GESTANT']=2

print(df['CS_GESTANT'].unique())

# 'DOSE_1_COV', 'DOSE_2_COV'

#Se tiver valor transformar para 1 
df['DOSE_1_COV'].isnull().value_counts()

df['DOSE_2_COV'].isnull().value_counts()

#Transforma colunas não nulas em 1 
df.loc[df['DOSE_1_COV'].isnull()==False,'DOSE_1_COV']=1
df.loc[df['DOSE_2_COV'].isnull()==False,'DOSE_2_COV']=1
df.loc[df['DOSE_1_COV'].isnull()==True,'DOSE_1_COV']=0
df.loc[df['DOSE_2_COV'].isnull()==True,'DOSE_2_COV']=0

df['DOSE_1_COV'].unique()

df['DOSE_2_COV'].unique()

# Mapear os dados: 

for cols in ColumnsObject:
    print(cols)
    print(df[cols].unique())
    print("------------------")
    print('\n')


mapa = {1:'Sim', 2:'Não', 9:'Ignorado',
        '1':'Sim', '2':'Não', '9':'Ignorado',
       1.0:'Sim', 2.0:'Não', 9.0:'Ignorado',
       }

mapa_suport = {1:'Invasivo', 2:'NãoInvasivo', 3:'Não'} 

cols1 = ['CS_GESTANT', 
         'PUERPERA',   
         'CARDIOPATI',
         'HEMATOLOGI',
         'SIND_DOWN',
         'HEPATICA',
         'ASMA',
         'DIABETES',
         'NEUROLOGIC',
         'PNEUMOPATI',
         'IMUNODEPRE',
         'RENAL',
         'OBESIDADE',   
        'NOSOCOMIAL',
        'FATOR_RISC',
         'FEBRE',
        'TOSSE',
        'GARGANTA',
        'DISPNEIA',
        'DESC_RESP',
        'SATURACAO',
        'DIARREIA',
        'VOMITO',
        'DOR_ABD',
        'FADIGA',
        'PERD_OLFT',
        'PERD_PALA', 
        'VACINA',
        'ANTIVIRAL',
        'SUPORT_VEN',
        'VACINA_COV']

for cols in cols1:
  if cols == 'SUPORT_VEN':
    df[cols]=df[cols].map(mapa_suport)
  else:
    df[cols]=df[cols].map(mapa)
    print(cols)
    print(df[cols].unique())
    print("------------------")
    print('\n')

#Na precisa fazer o mapper aqui, ja foi resolvido na celula acima
cols = [ 'CARDIOPATI',
        'HEMATOLOGI',
        'SIND_DOWN',
        'HEPATICA',
        'ASMA',
        'DIABETES',
        'NEUROLOGIC',
        'PNEUMOPATI',
        'IMUNODEPRE',
        'RENAL',
        'OBESIDADE']
for col in cols:
    print(col)
    print(df[col].unique())
    print("------------------")
    print('\n')



#Realizado a substituicao da numeracao para descricao
df['CS_RACA'] = df['CS_RACA'].map({1:'Branca', 2:'Preta',3:'Amarela',4:'Parda',5:'Indigena',9:'Ignorado'})

df.head()

#Removendo pessoas que tomaram a dose 2 e não tomaram a Dose 1 da vacina
df = df[~((df['DOSE_1_COV'] == 0) & (df['DOSE_2_COV'] == 1))]

#Separando as colunas que serão utilizadas no modelo
maincols = ['CS_SEXO', 'CS_GESTANT', 'CS_RACA',
       'NOSOCOMIAL', 'FEBRE', 'TOSSE', 'GARGANTA', 'DISPNEIA', 'DESC_RESP',
       'SATURACAO', 'DIARREIA', 'VOMITO', 'PUERPERA', 'FATOR_RISC',
       'CARDIOPATI', 'HEMATOLOGI', 'SIND_DOWN', 'HEPATICA', 'ASMA', 'DIABETES',
       'NEUROLOGIC', 'PNEUMOPATI', 'IMUNODEPRE', 'RENAL', 'OBESIDADE',
       'VACINA', 'ANTIVIRAL', 'SUPORT_VEN', 'DOR_ABD', 'FADIGA',
       'PERD_OLFT', 'PERD_PALA', 'VACINA_COV', 'DOSE_1_COV', 'DOSE_2_COV']

df = df[maincols]

#Retirando as colunas com inconsistencia  
df = df[~(((df['DOSE_1_COV'] == 1)) & ((df['VACINA_COV'] == 'Não') | (df['VACINA_COV'] == 'Ignorado')))]

#Retirando as colunas com inconsistencia  
df = df[~(((df['DOSE_2_COV'] == 1)) & ((df['VACINA_COV'] == 'Não') | (df['VACINA_COV'] == 'Ignorado')))]

df.shape

maincols = ['CS_SEXO', 'CS_GESTANT', 'CS_RACA',
       'NOSOCOMIAL', 'FEBRE', 'TOSSE', 'GARGANTA', 'DISPNEIA', 'DESC_RESP',
       'SATURACAO', 'DIARREIA', 'VOMITO', 'PUERPERA', 'FATOR_RISC',
       'CARDIOPATI', 'HEMATOLOGI', 'SIND_DOWN', 'HEPATICA', 'ASMA', 'DIABETES',
       'NEUROLOGIC', 'PNEUMOPATI', 'IMUNODEPRE', 'RENAL', 'OBESIDADE',
       'VACINA', 'ANTIVIRAL', 'SUPORT_VEN', 'DOR_ABD', 'FADIGA',
       'PERD_OLFT', 'PERD_PALA', 'DOSE_1_COV', 'DOSE_2_COV']

df = df[maincols]

df.head()

df.to_pickle('INFLUD21-27-09-2021-Tratado.pkl')
