==> modelagem/04 - GetDummies.ipynb <==
import pandas as pd 
import numpy as np
from google.colab import drive
from datetime import datetime
drive.mount('/content/drive')

df = pd.read_pickle('/content/drive/MyDrive/Grupo Facens DataScience/TCC/Grupo 1/Codes/Arquivos/INFLUD21-27-09-2021-Tratado.pkl')

df.shape

df.columns

df.head()

df_dummies = df.drop(['SUPORT_VEN','FATOR_RISC'], axis=1)

lista_carac_2 = []
lista_carac_mais = []

for col in df_dummies.columns:
  if len(df_dummies[col].unique()) > 2:
    lista_carac_mais.append(col)
  else:
    lista_carac_2.append(col)

df_dummies = pd.get_dummies(df_dummies[lista_carac_mais])

for col in lista_carac_2:
  df_dummies[col] = df[col]

df_dummies['SUPORT_VEN'] = df['SUPORT_VEN'].map({'Invasivo': 1, 'NãoInvasivo': 2, 'Não': 3})

Realziar o tratamento, mesclando algumas informacoes como por exemplo 
------------------------------------
Invasivo: 1, NãoInvasivo: 2, = SIM
Não: 3 = NAO 
------------------------------------
Invasivo: 1,  = SIM
Não: 3, NãoInvasivo: 2 = NAO 
------------------------------------
Invasivo: 1 = SIM 1
NãoInvasivo: 2 = SIM 2
Não: 3 = NAO 
------------------------------------

Posteriormente gerar 3 Pickles

df_dummies.head()

###V1
#Invasivo: 1 = SIM 1
#NãoInvasivo: 2 = SIM 2
#Não: 3 = NAO 

df_dummies.to_pickle('INFLUD21-27-09-2021-TratadoWDummies.pkl')

#V2
#Invasivo: 1,  = SIM
#Não: 3, NãoInvasivo: 2 = NAO 

df_dummiesInv1_NaoInv2_Nao.to_pickle('INFLUD21-27-09-2021-TratadoWDummies_Inv1_NaoInv2_Nao2.pkl')
