==> modelagem/4.5 - Avaliação do Modelo MLJAR.ipynb <==
!pip install mljar-supervised

pip install pandas --upgrade

import pandas as pd 
import numpy as np
from google.colab import drive
from datetime import datetime
from supervised.automl import AutoML
from supervised.preprocessing.eda import EDA
from sklearn.model_selection import train_test_split
drive.mount('/content/drive')

df = pd.read_pickle('/content/drive/MyDrive/Grupo Facens DataScience/TCC/Grupo 1/Codes/Arquivos/INFLUD21-27-09-2021-TratadoWDummies.pkl')

y = df['SUPORT_VEN'] 
x = df.drop(['SUPORT_VEN'],axis=1)

X_train, X_test, y_train, y_test = train_test_split(x, y, test_size=0.33, random_state=42)

total_time_limit = 60
automl = AutoML(total_time_limit=total_time_limit)
automl.fit(X_train,y_train)
automl.report()
